google auth almost done
returns the user info, using hardcoded valuse for now to bypass the problem with node env undefined. disabled strict mode so that the app doesnt pass two requests to google api and ultimately return error instead of a correct response. need to implement the creation of a new user with the retrieved values